Abstract out the external session cache
The nginx codebase currently assumes that you've either got an
OpenSSL-builtin session cache, or a SHM-based shared session cache (which it
calls the "external session cache").  Since we're trying to add a second
form of "external session cache" (memcached), we need to start by avoiding
referring directly to the SHM cache as our external session cache, and put
in an abstract middleman that will allow us to refer to "the external
session cache" in general to anyone who doesn't need to know exactly what
session cache we're using.  For now, that just means adding another struct
that stores the shm_zone and using that everywhere.

Next step: adding a reference to a memcached store type.