Remove keys from memcache if requested, and cleanup error handling
Removing keys is easy -- a quick mc_delete and we're done.  The error
handling was basically just moving the big blob of end-times cleanup at the
end of each function into per-store-type blobs.  This means that errors in
(say) the SHM store won't prevent a memcache store request from succeeding
(in the very unlikely event that both are running).